pyproject.toml
[project]
name = "balaambot"
version = "0.1.0"
description = "A discord bot for dutty beats"
authors = [
    { name = "James Wild", email = "wild.james343@gmail.com" },
    { name = "Jibby", email = "Jibbajabbafic@users.noreply.github.com" },
]
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "discord.py>=2.5.2",
    "anyio>=4.9.0",
    "dotenv>=0.9.9",
    "pynacl>=1.5.0",
    "discord-ext-voice-recv>=0.4.2a145",
    "yt-dlp>=2025.5.22",
    "pytest-cov>=6.1.1",
    "pyjokes>=0.8.3",
    "pydantic>=2.11.5",
    "redis>=6.2.0",
    "pytest-asyncio>=1.0.0",
]

# Dev tools
[dependency-groups]
dev = [
    "ipdb>=0.13.13",
    "ipython>=8.37.0",
    "pre-commit>=4.1.0",
    "pyright>=1.1.392.post0",
    "pytest>=8.3.4",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.2",
]

[project.scripts]
balaambot = "balaambot:main.start"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
